refactor(worker): extract workers/http.mjs response primitives from api.mjs (#510)
Third slice of the api.mjs de-monolith — the shared-util keystone. Move the
three pure HTTP response primitives apiHeaders / errorResponse / weakEtag
(74 combined call sites) into a new leaf module workers/http.mjs.

Pure restructuring — NO behaviour change. http.mjs imports only CACHE_SECONDS +
CONTRACT_VERSION (../src/contracts.mjs) and JSON_CONTENT_TYPE (./config.mjs);
nothing from api.mjs, so no cycle. The more entangled response builders
(envelopeResponse, dataResponse, corsPreflight) stay in api.mjs since they read
env/pointer data or need WEBHOOK_SECRET_HEADER — they now consume the imported
primitives. The three constants remain used elsewhere in api.mjs, so their
imports stay.

This unblocks the upcoming handler-group splits (rpc, badge, analytics, …),
which all depend on these primitives.

api.mjs: 3301 -> 3248 LOC. None of the three are exported / imported by tests
(exercised via handleRequest), so no re-exports.

Verified: 122 worker tests pass (worker-runtime, api-coverage, webhooks-worker,
subnet-overview), worker:test (esbuild) green, validate:contract-drift clean,
lint clean. No logic diff.
